Make it clear we're wating, not retrying
During usability testing we asked them to push a docker image.

The first two participants were confused by ths terminal output:

    0 of 1 instances running, 1 starting
    0 of 1 instances running, 1 starting
    0 of 1 instances running, 1 starting

Participants thought that an instance was being started in a loop and each line
printed corresponded to an attempt to start the app. Both of them thought that
this message meant that the push of the docker image was likely to fail and
they should cancel it.

In fact, this message can be printed a number of times prior to a container
being successfully started.

This is a temporary commit that tries to fix this by printing an opening line
and then a dot each time the cli checks how many instances there are. We'll get
the next participants to try this.